requirements.txt
numpy==2.3.5
pandas==2.3.3
scikit-learn==1.7.2
matplotlib==3.10.7
pillow==11.3.0
tqdm==4.67.1
ipykernel==6.31.0
torchinfo==1.8.0
wandb==0.23.1
dotenv==0.9.9